BREAKING CHANGES: structureSearch helper
A lot of code has been removed because it is now includes in the
openchemlib-extended library.

This library is now able to deal with 'flatten' result (the result
is an array of entries) or an array containing a chemical structure
linked to 'data' that is an array of entries.

To create the database here is an example of code based on a Toc:

const db = await structureSearch.buildDatabase(data, {
   showLoading: true
});

To search the database the code will look like:

const options = {
    mode: 'substructure',  // or 'exact' or 'similarity'
    format: 'molfile',  // or 'smiles' or 'idcode'
    flattenResult: false  // by default the result if flatten
};
let result = db.search(molfile, options);